+++ src/routes/+layout.svelte
@@ -11,12 +11,13 @@
 	$: server = $page.data.serverInfo || {};
 	/** @type {'v1'|'v2'} */
 	$: apiVersion = $page.url.pathname.startsWith('/v1') ? 'v1' : 'v2';
+	$: warningBanner = $page.data.warningBanner;
 	// @ts-ignore
 	// eslint-disable-next-line no-undef
 	let clientVersion = __APP_VERSION__;
 
 	$: displayVersionSelector =
-		$page.data.v1Enabled &&
+		$page.data.apiV1Mode !== 'exclude' &&
 		(!isSubPage($page.url.pathname, apiVersion) ||
 			$page.url.pathname === '/v2/admin/jobs' ||
 			$page.url.pathname === '/v1/admin/jobs') &&
@@ -218,6 +219,19 @@
 		{#if !server.alive}
 			<div class="alert alert-danger">
 				Sorry, we are performing some maintenance on fractal-server. It will be back online soon.
+			</div>
+		{/if}
+		{#if apiVersion === 'v1' && $page.data.apiV1Mode === 'include_read_only'}
+			<div class="alert alert-warning">Warning: legacy API is in read-only mode.</div>
+		{/if}
+		{#if warningBanner}
+			<div class="alert alert-warning">
+				{#each warningBanner.split('\n') as line, index}
+					{#if index > 0}
+						<br />
+					{/if}
+					{line}
+				{/each}
 			</div>
 		{/if}
 		{#if userLoggedIn && !$page.data.userInfo.is_verified}